Added Makefile and fixed errors stopping 'make' from running correctly.

+++ LoadBalancer.h
@@ -9,7 +9,7 @@ using namespace std;
 
 class LoadBalancer {
 public:
-    LoadBalancer(int maxNumServers);
+    LoadBalancer(size_t maxNumServers);
     ~LoadBalancer();
     void distributeReq(const Request& request, int currTime); //gives a request to a server to process only if it is available or adds it to the main queue
     void simTime(int currTime); //Simulates time step in the system, updates servers and processes queued requests
@@ -23,12 +23,12 @@ private:
     vector<WebServer*> notRunningServers; //Vector of idle servers
     queue<Request> reqQueue;
 
-    int maxNumServers; //Maximum amount of servers allowed to be activated
-    int minNumServers; //Minimum number of servers allowed to be activated (will automatically start up upon running the code)
-    int nextServerID; 
+    size_t maxNumServers; //Maximum amount of servers allowed to be activated
+    size_t minNumServers; //Minimum number of servers allowed to be activated (will automatically start up upon running the code)
+    size_t nextServerID; 
     
-    int idleServerCountToRemove; //How many idle servers before removing any
-    int reqQueueSizeToAdd; //How many requests in queue to trigger the activation of a new server
+    size_t idleServerCountToRemove; //How many idle servers before removing any
+    size_t reqQueueSizeToAdd; //How many requests in queue to trigger the activation of a new server
 
     void procQueuedReqs(int currTime); //Processes queued requests by giving them to a server
     void addServer(int currTime);
